requirements.txt
requests
pandas
geopandas
shapely
pyyaml
tqdm
python-dotenv
rasterio
pyproj
fiona